new implementation for the combat interaction
the problem was that, depending on many things, combat interaction can change in many ways, and can be called recursively.
To manage that, there is a state object, which accumulates changes and allows next invokations to know about what happened previously.
And basing on this, making further decisions.
Some sort of 'Finite state machine', kind of thing